+++ src/tool/main.rs
@@ -25,6 +25,7 @@ where
                 Subcommand::Snapshot(cmd) => cmd.run().await,
                 Subcommand::Fetch(cmd) => cmd.run().await,
                 Subcommand::Archive(cmd) => cmd.run().await,
+                Subcommand::DB(cmd) => cmd.run().await,
             }
         })
 }